Revert "Merge pull request #4209 from mayarub/87123-enable-vsock-maya"
This reverts commit a759a4923aadfa40f2d2520bf6235f0f93d865eb, reversing
changes made to ab1e8c01fc6170dbede842cef7cda139d9c3c1d7.
