+++ locales/en/plugin__kubevirt-plugin.json
@@ -283,7 +283,6 @@
   "Attach this data to a VirtualMachine operating system": "Attach this data to a VirtualMachine operating system",
   "Attach VirtualMachine to multiple networks": "Attach VirtualMachine to multiple networks",
   "Attached NVIDIA GPU resources": "Attached NVIDIA GPU resources",
-  "Attaches a virtio-vsock device for direct communication between the guest and the host without a network interface.": "Attaches a virtio-vsock device for direct communication between the guest and the host without a network interface.",
   "Auto converge": "Auto converge",
   "Auto update": "Auto update",
   "Auto-attached": "Auto-attached",
@@ -821,7 +820,6 @@
   "Enable predictive analytics": "Enable predictive analytics",
   "Enable the \"Enable folders in VirtualMachines tree view\" preview feature in Settings > Preview features to use folders.": "Enable the \"Enable folders in VirtualMachines tree view\" preview feature in Settings > Preview features to use folders.",
   "Enable the creation of LoadBalancer services for SSH connections to VirtualMachines. A load balancer must be configured": "Enable the creation of LoadBalancer services for SSH connections to VirtualMachines. A load balancer must be configured",
-  "Enable VSOCK": "Enable VSOCK",
   "Enable Windows Server 2022 golden image creation and testing. Requires accepting the Windows End User License Agreement (EULA).": "Enable Windows Server 2022 golden image creation and testing. Requires accepting the Windows End User License Agreement (EULA).",
   "Enabled": "Enabled",
   "Enables access to the VirtualMachine's guest system log. Wait a few seconds for logging to start before viewing the log.": "Enables access to the VirtualMachine's guest system log. Wait a few seconds for logging to start before viewing the log.",
