+++ locales/es/plugin__kubevirt-plugin.json
@@ -294,7 +294,6 @@
   "Attach this data to a VirtualMachine operating system": "Attach this data to a VirtualMachine operating system",
   "Attach VirtualMachine to multiple networks": "Adjunte la VirtualMachine a múltiples redes",
   "Attached NVIDIA GPU resources": "Adjunte recursos GPU de NVIDIA",
-  "Attaches a virtio-vsock device for direct communication between the guest and the host without a network interface.": "Attaches a virtio-vsock device for direct communication between the guest and the host without a network interface.",
   "Auto converge": "Convergencia automática",
   "Auto update": "Actualización automática",
   "Auto-attached": "Se acopla automáticamente",
@@ -835,7 +834,6 @@
   "Enable predictive analytics": "Habilitar el análisis predictivo",
   "Enable the \"Enable folders in VirtualMachines tree view\" preview feature in Settings > Preview features to use folders.": "Enable the \"Enable folders in VirtualMachines tree view\" preview feature in Settings > Preview features to use folders.",
   "Enable the creation of LoadBalancer services for SSH connections to VirtualMachines. A load balancer must be configured": "Habilite la creación de servicios de LoadBalancer para conexiones SSH a VirtualMachines. Se debe configurar un equilibrador de carga.",
-  "Enable VSOCK": "Enable VSOCK",
   "Enable Windows Server 2022 golden image creation and testing. Requires accepting the Windows End User License Agreement (EULA).": "Enable Windows Server 2022 golden image creation and testing. Requires accepting the Windows End User License Agreement (EULA).",
   "Enabled": "Activado",
   "Enables access to the VirtualMachine's guest system log. Wait a few seconds for logging to start before viewing the log.": "Permite acceder al registro del sistema invitado de la VirtualMachine. Espere unos segundos a que comience el registro antes de verlo.",
