+++ locales/fr/plugin__kubevirt-plugin.json
@@ -294,7 +294,6 @@
   "Attach this data to a VirtualMachine operating system": "Attach this data to a VirtualMachine operating system",
   "Attach VirtualMachine to multiple networks": "Attacher une machine virtuelle à plusieurs réseaux",
   "Attached NVIDIA GPU resources": "Ressources GPU NVIDIA jointes",
-  "Attaches a virtio-vsock device for direct communication between the guest and the host without a network interface.": "Attaches a virtio-vsock device for direct communication between the guest and the host without a network interface.",
   "Auto converge": "Convergence automatique",
   "Auto update": "Mise à jour automatique",
   "Auto-attached": "Attachement automatique",
@@ -835,7 +834,6 @@
   "Enable predictive analytics": "Activer l'analyse prédictive",
   "Enable the \"Enable folders in VirtualMachines tree view\" preview feature in Settings > Preview features to use folders.": "Enable the \"Enable folders in VirtualMachines tree view\" preview feature in Settings > Preview features to use folders.",
   "Enable the creation of LoadBalancer services for SSH connections to VirtualMachines. A load balancer must be configured": "Activez la création de services LoadBalancer pour les connexions SSH aux machines virtuelles. Un équilibreur de charge doit être configuré.",
-  "Enable VSOCK": "Enable VSOCK",
   "Enable Windows Server 2022 golden image creation and testing. Requires accepting the Windows End User License Agreement (EULA).": "Enable Windows Server 2022 golden image creation and testing. Requires accepting the Windows End User License Agreement (EULA).",
   "Enabled": "Activé",
   "Enables access to the VirtualMachine's guest system log. Wait a few seconds for logging to start before viewing the log.": "Permet d'accéder au journal système invité de la machine virtuelle. Attendez quelques secondes que la journalisation démarre avant de l'afficher.",
