+++ locales/ja/plugin__kubevirt-plugin.json
@@ -281,7 +281,6 @@
   "Attach this data to a VirtualMachine operating system": "Attach this data to a VirtualMachine operating system",
   "Attach VirtualMachine to multiple networks": "VirtualMachine を複数のネットワークに割り当てます",
   "Attached NVIDIA GPU resources": "割り当てられた NVIDIA GPU リソース",
-  "Attaches a virtio-vsock device for direct communication between the guest and the host without a network interface.": "Attaches a virtio-vsock device for direct communication between the guest and the host without a network interface.",
   "Auto converge": "自動収束",
   "Auto update": "自動更新",
   "Auto-attached": "自動アタッチ",
@@ -818,7 +817,6 @@
   "Enable predictive analytics": "予測分析を有効にする",
   "Enable the \"Enable folders in VirtualMachines tree view\" preview feature in Settings > Preview features to use folders.": "Enable the \"Enable folders in VirtualMachines tree view\" preview feature in Settings > Preview features to use folders.",
   "Enable the creation of LoadBalancer services for SSH connections to VirtualMachines. A load balancer must be configured": "VirtualMachine に SSH 接続するための LoadBalancer サービスの作成を有効にします。ロードバランサーを設定する必要があります",
-  "Enable VSOCK": "Enable VSOCK",
   "Enable Windows Server 2022 golden image creation and testing. Requires accepting the Windows End User License Agreement (EULA).": "Enable Windows Server 2022 golden image creation and testing. Requires accepting the Windows End User License Agreement (EULA).",
   "Enabled": "有効",
   "Enables access to the VirtualMachine's guest system log. Wait a few seconds for logging to start before viewing the log.": "VirtualMachine のゲストシステムログへのアクセスを有効にします。ログを表示する前に、ロギングが開始するまで数秒お待ちください。",
