+++ locales/ko/plugin__kubevirt-plugin.json
@@ -281,7 +281,6 @@
   "Attach this data to a VirtualMachine operating system": "Attach this data to a VirtualMachine operating system",
   "Attach VirtualMachine to multiple networks": "여러 네트워크에 VirtualMachine 연결",
   "Attached NVIDIA GPU resources": "연결된 NVIDIA GPU 리소스",
-  "Attaches a virtio-vsock device for direct communication between the guest and the host without a network interface.": "Attaches a virtio-vsock device for direct communication between the guest and the host without a network interface.",
   "Auto converge": "자동 통합",
   "Auto update": "자동 업데이트",
   "Auto-attached": "자동으로 연결됨",
@@ -818,7 +817,6 @@
   "Enable predictive analytics": "예측 분석 활성화",
   "Enable the \"Enable folders in VirtualMachines tree view\" preview feature in Settings > Preview features to use folders.": "Enable the \"Enable folders in VirtualMachines tree view\" preview feature in Settings > Preview features to use folders.",
   "Enable the creation of LoadBalancer services for SSH connections to VirtualMachines. A load balancer must be configured": "VirtualMachines에 대한 SSH 연결을 위해 LoadBalancer 서비스 생성을 활성화합니다. 로드 밸런서는 별도로 구성되어 있어야 합니다.",
-  "Enable VSOCK": "Enable VSOCK",
   "Enable Windows Server 2022 golden image creation and testing. Requires accepting the Windows End User License Agreement (EULA).": "Enable Windows Server 2022 golden image creation and testing. Requires accepting the Windows End User License Agreement (EULA).",
   "Enabled": "활성화됨",
   "Enables access to the VirtualMachine's guest system log. Wait a few seconds for logging to start before viewing the log.": "VirtualMachine의 게스트 시스템 로그에 대한 액세스를 활성화합니다. 로그를 보기 전에 로깅이 시작될 때까지 몇 초 동안 기다립니다.",
