+++ locales/zh/plugin__kubevirt-plugin.json
@@ -281,7 +281,6 @@
   "Attach this data to a VirtualMachine operating system": "Attach this data to a VirtualMachine operating system",
   "Attach VirtualMachine to multiple networks": "将虚拟机附加到多个网络",
   "Attached NVIDIA GPU resources": "附加的 NVIDIA GPU 资源",
-  "Attaches a virtio-vsock device for direct communication between the guest and the host without a network interface.": "Attaches a virtio-vsock device for direct communication between the guest and the host without a network interface.",
   "Auto converge": "自动聚合",
   "Auto update": "自动更新",
   "Auto-attached": "自动附加",
@@ -818,7 +817,6 @@
   "Enable predictive analytics": "启用预测分析",
   "Enable the \"Enable folders in VirtualMachines tree view\" preview feature in Settings > Preview features to use folders.": "Enable the \"Enable folders in VirtualMachines tree view\" preview feature in Settings > Preview features to use folders.",
   "Enable the creation of LoadBalancer services for SSH connections to VirtualMachines. A load balancer must be configured": "为到 VirtualMachines 的 SSH 连接启用 LoadBalancer 服务创建。必须配置一个负载均衡器",
-  "Enable VSOCK": "Enable VSOCK",
   "Enable Windows Server 2022 golden image creation and testing. Requires accepting the Windows End User License Agreement (EULA).": "Enable Windows Server 2022 golden image creation and testing. Requires accepting the Windows End User License Agreement (EULA).",
   "Enabled": "已启用",
   "Enables access to the VirtualMachine's guest system log. Wait a few seconds for logging to start before viewing the log.": "启用对 VirtualMachine 客户机系统日志的访问。等待几秒钟以便日志记录在查看日志前启动。",
