+++ src/utils/hooks/useVMTemplateFeatureFlag/constants.ts
@@ -1,12 +1,9 @@
-import { Patch } from '@openshift-console/dynamic-plugin-sdk';
-
-export { KUBEVIRT_JSONPATCH_ANNOTATION } from '@kubevirt-utils/hooks/utils/featureGateAnnotation';
-
+export const KUBEVIRT_JSONPATCH_ANNOTATION = 'kubevirt.kubevirt.io/jsonpatch';
 export const TEMPLATE_FEATURE_GATE = 'Template';
 export const TEMPLATE_FEATURE_GATE_PATH =
   '/spec/configuration/developerConfiguration/featureGates/-';
 
-export const ADD_TEMPLATE_FEATURE_GATE_PATCH: Patch = {
+export const ADD_TEMPLATE_FEATURE_GATE_PATCH = {
   op: 'add',
   path: TEMPLATE_FEATURE_GATE_PATH,
   value: TEMPLATE_FEATURE_GATE,
